Ready for submission
Comments added for integration of audio.

Fixed warning caused by unreferenced parameter, implicit conversions of variable etc..

+++ Artless/ALEngine/Source/Editor/Panels/ContentBrowserPanel.cpp
@@ -164,7 +164,7 @@ namespace ALEngine::Editor
 			if (fileNamestring.find(".jpg")!= std::string::npos || fileNamestring.find(".png") != std::string::npos)
 			{
 				Guid id = Engine::AssetManager::Instance()->GetGuid(directoryEntry.path().string());
-				u32 texture = Engine::AssetManager::Instance()->GetButtonImage(id);
+				u64 texture = static_cast<u64>(Engine::AssetManager::Instance()->GetButtonImage(id));
 				ImGui::ImageButton(reinterpret_cast<ImTextureID>(texture), { thumbnailSize, thumbnailSize }, { 0, 1 }, { 1, 0 });
 			}
 			else
